Idiomatically organize pruning DB tests
