+++ core/lib/dal/src/lib.rs
@@ -35,6 +35,7 @@ pub mod eth_sender_dal;
 pub mod events_dal;
 pub mod events_web3_dal;
 pub mod factory_deps_dal;
+pub mod metrics;
 mod models;
 pub mod proof_generation_dal;
 pub mod protocol_versions_dal;
@@ -54,13 +55,8 @@ pub mod tokens_web3_dal;
 pub mod transactions_dal;
 pub mod transactions_web3_dal;
 
-pub mod metrics;
-
 #[cfg(test)]
 mod tests;
-
-#[cfg(test)]
-mod pruning_dal_tests;
 
 // This module is private and serves as a way to seal the trait.
 mod private {
